rename to cfgmetadatagen
