+++ cmd/configschema/cfgmetadatagen/cfgmetadatagen/source_tree_yaml_writer.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package yamlgen
+package cfgmetadatagen
 
 import (
 	"fmt"
